post_service/requirements.txt
pillow
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
drf-spectacular